add oferta educativa y tutor

+++ src/main/java/com/amdocs/jshapi/repository/BenefactorRepository.java
@@ -9,7 +9,7 @@ import com.amdocs.jshapi.domain.Region;
 import com.amdocs.jshapi.projections.BenefactorInline;
 
 @RepositoryRestResource(path = "benefactores", collectionResourceRel = "benefactores", excerptProjection = BenefactorInline.class)
-public interface BenefactorRepository extends PagingAndSortingRepository<Benefactor, Integer>{
+public interface BenefactorRepository extends PagingAndSortingRepository<Benefactor, Long>{
 	
 	public Benefactor findBenefactorByNombre(@Param("nombre") String nombre);
 	//public Benefactor findBenefactorByTipo(@Param("tipobeneficiario") int tipobeneficiario);
